fix: changed int type to string of blog_id
