--- alembic/versions/1bd44db2d5c5_added_jobs_table.py
@@ -1,46 +1,0 @@
-"""added jobs table
-
-Revision ID: 1bd44db2d5c5
-Revises: 782e99b1b3dc
-Create Date: 2024-07-24 07:19:38.935749
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '1bd44db2d5c5'
-down_revision: Union[str, None] = '782e99b1b3dc'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('jobs',
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('title', sa.Text(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=False),
-    sa.Column('department', sa.String(), nullable=True),
-    sa.Column('location', sa.String(), nullable=True),
-    sa.Column('salary', sa.String(), nullable=True),
-    sa.Column('job_type', sa.String(), nullable=True),
-    sa.Column('company_name', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_jobs_id'), 'jobs', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_jobs_id'), table_name='jobs')
-    op.drop_table('jobs')
-    # ### end Alembic commands ###
