--- alembic/versions/22e29cebfcb5_added_oauth_table.py
@@ -1,30 +1,0 @@
-"""added oauth table
-
-Revision ID: 22e29cebfcb5
-Revises: b86c35b90b4c
-Create Date: 2024-07-24 07:25:46.813821
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '22e29cebfcb5'
-down_revision: Union[str, None] = 'b86c35b90b4c'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
