--- alembic/versions/47456f2e62a3_added_blogs_table.py
@@ -1,30 +1,0 @@
-"""added blogs table
-
-Revision ID: 47456f2e62a3
-Revises: d1058cc4e645
-Create Date: 2024-07-24 07:41:54.219711
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '47456f2e62a3'
-down_revision: Union[str, None] = 'd1058cc4e645'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
