--- alembic/versions/782e99b1b3dc_added_activity_logs_table.py
@@ -1,30 +1,0 @@
-"""added activity logs table
-
-Revision ID: 782e99b1b3dc
-Revises: 4871c2a01ae2
-Create Date: 2024-07-24 07:15:53.837544
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '782e99b1b3dc'
-down_revision: Union[str, None] = '4871c2a01ae2'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
