--- alembic/versions/b1aa146bed47_added_profile_table.py
@@ -1,65 +1,0 @@
-"""added profile table
-
-Revision ID: b1aa146bed47
-Revises: ced363af8b02
-Create Date: 2024-07-24 07:03:37.378653
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b1aa146bed47'
-down_revision: Union[str, None] = 'ced363af8b02'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('org_roles',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('org_id', sa.String(), nullable=False),
-    sa.Column('is_admin', sa.Boolean(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['org_id'], ['organizations.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_org_roles_id'), 'org_roles', ['id'], unique=False)
-    op.create_table('profiles',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('username', sa.String(), nullable=True),
-    sa.Column('first_name', sa.String(), nullable=True),
-    sa.Column('last_name', sa.String(), nullable=True),
-    sa.Column('pronouns', sa.String(), nullable=True),
-    sa.Column('job_title', sa.String(), nullable=True),
-    sa.Column('department', sa.String(), nullable=True),
-    sa.Column('social', sa.Text(), nullable=True),
-    sa.Column('bio', sa.Text(), nullable=True),
-    sa.Column('phone_number', sa.String(), nullable=True),
-    sa.Column('avatar_url', sa.String(), nullable=True),
-    sa.Column('recovery_email', sa.String(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_profiles_id'), 'profiles', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_profiles_id'), table_name='profiles')
-    op.drop_table('profiles')
-    op.drop_index(op.f('ix_org_roles_id'), table_name='org_roles')
-    op.drop_table('org_roles')
-    # ### end Alembic commands ###
