--- alembic/versions/b86c35b90b4c_added_token_login_table.py
@@ -1,30 +1,0 @@
-"""added token login table
-
-Revision ID: b86c35b90b4c
-Revises: 1bd44db2d5c5
-Create Date: 2024-07-24 07:23:28.618828
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b86c35b90b4c'
-down_revision: Union[str, None] = '1bd44db2d5c5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
