--- alembic/versions/ced363af8b02_added_org_role_tables.py
@@ -1,30 +1,0 @@
-"""added org_role tables
-
-Revision ID: ced363af8b02
-Revises: 3d7695723bfa
-Create Date: 2024-07-24 06:55:50.835200
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'ced363af8b02'
-down_revision: Union[str, None] = '3d7695723bfa'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
