--- alembic/versions/d1058cc4e645_added_payments_table.py
@@ -1,30 +1,0 @@
-"""added payments table
-
-Revision ID: d1058cc4e645
-Revises: f962c00d62eb
-Create Date: 2024-07-24 07:34:48.866394
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd1058cc4e645'
-down_revision: Union[str, None] = 'f962c00d62eb'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
