--- alembic/versions/d5955ee5de29_added_comments_table.py
@@ -1,30 +1,0 @@
-"""added comments table
-
-Revision ID: d5955ee5de29
-Revises: 47456f2e62a3
-Create Date: 2024-07-24 07:46:30.492088
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd5955ee5de29'
-down_revision: Union[str, None] = '47456f2e62a3'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
