--- alembic/versions/d68f9b94e9a3_added_billing_plans_tables.py
@@ -1,30 +1,0 @@
-"""added billing plans tables
-
-Revision ID: d68f9b94e9a3
-Revises: d5955ee5de29
-Create Date: 2024-07-24 07:50:50.821158
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'd68f9b94e9a3'
-down_revision: Union[str, None] = 'd5955ee5de29'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
