--- alembic/versions/f962c00d62eb_added_testimonial_table.py
@@ -1,74 +1,0 @@
-"""added testimonial table
-
-Revision ID: f962c00d62eb
-Revises: 22e29cebfcb5
-Create Date: 2024-07-24 07:30:15.255053
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f962c00d62eb'
-down_revision: Union[str, None] = '22e29cebfcb5'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('oauth',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('provider', sa.String(), nullable=False),
-    sa.Column('sub', sa.String(), nullable=False),
-    sa.Column('access_token', sa.String(), nullable=False),
-    sa.Column('refresh_token', sa.String(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_oauth_id'), 'oauth', ['id'], unique=False)
-    op.create_table('testimonials',
-    sa.Column('client_designation', sa.String(), nullable=True),
-    sa.Column('client_name', sa.String(), nullable=True),
-    sa.Column('author_id', sa.String(), nullable=False),
-    sa.Column('comments', sa.Text(), nullable=True),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('ratings', sa.Float(), nullable=True),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['author_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_testimonials_id'), 'testimonials', ['id'], unique=False)
-    op.create_table('token_logins',
-    sa.Column('user_id', sa.String(), nullable=False),
-    sa.Column('token', sa.String(), nullable=False),
-    sa.Column('expiry_time', sa.DateTime(), nullable=False),
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id')
-    )
-    op.create_index(op.f('ix_token_logins_id'), 'token_logins', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_token_logins_id'), table_name='token_logins')
-    op.drop_table('token_logins')
-    op.drop_index(op.f('ix_testimonials_id'), table_name='testimonials')
-    op.drop_table('testimonials')
-    op.drop_index(op.f('ix_oauth_id'), table_name='oauth')
-    op.drop_table('oauth')
-    # ### end Alembic commands ###
