+++ api/v1/models/__init__.py
@@ -1,16 +1,20 @@
+from api.v1.models.activity_logs import ActivityLog
+from api.v1.models.billing_plan import BillingPlan
+from api.v1.models.comment import Comment
+from api.v1.models.contact_us import ContactUs
+from api.v1.models.message import Message
+from api.v1.models.payment import Payment
+from api.v1.models.waitlist import Waitlist
 from api.v1.models.user import User
 from api.v1.models.org import Organization
 from api.v1.models.org_role import OrgRole
 from api.v1.models.profile import Profile
 from api.v1.models.notifications import Notification
-# from api.v1.models.product import Product
-# from api.v1.models.subscription import Subscription
-# from api.v1.models.blog import Blog
+from api.v1.models.product import Product
+from api.v1.models.blog import Blog
 from api.v1.models.job import Job
 from api.v1.models.testimonial import Testimonial
 from api.v1.models.token_login import TokenLogin
 from api.v1.models.oauth import OAuth
-# from api.v1.models.invitation import Invitation
-# from api.v1.models.role import Role
-# from api.v1.models.permission import Permission
-# from api.v1.models.newsletter import Newsletter
+from api.v1.models.invitation import Invitation
+from api.v1.models.newsletter import Newsletter
