+++ api/v1/models/comment.py
@@ -6,7 +6,7 @@ class Comment(BaseTableModel):
     __tablename__ = "comments"
 
     user_id = Column(String, ForeignKey('users.id', ondelete="CASCADE"), nullable=False)
-    blog_id = Column(Integer, ForeignKey('blogs.id', ondelete="CASCADE"), nullable=False)
+    blog_id = Column(String, ForeignKey('blogs.id', ondelete="CASCADE"), nullable=False)
     content = Column(Text, nullable=False)
 
     user = relationship("User", back_populates="comments")
